jipdate: fix regression with file operation
A regression was found in
49866d0 jipdate: avoid refetching issues during the parsing of the text file

This patch introduced an optimization when using a user query
(-q). However when using file mode, we don't run a Jira query in the
initial phase, so once we do the parsing, the issues have not been
fetched yet.

This patch fixes this situation, so now we do the following:

1. we check if we have fetched issues already
2. if so, we check if the issue we are looking for exist in the
already fetched set
3. otherwise, we query Jira

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>